refactor: implement EXCSRosterManager for managing Roster entries in EX-CommandStation
